refactor MPI  exchange routine.
In the future, will allow exchange routines not to
exchange normal field last active zone when shearing box is enabled.

For this, the refactoring include an optional parameter "overwriteBXn"
in each exchange direction
