+++ src/mpi/CMakeLists.txt
@@ -1,0 +1,7 @@
+target_sources(idefix
+  PUBLIC ${CMAKE_CURRENT_LIST_DIR}/buffer.hpp
+  PUBLIC ${CMAKE_CURRENT_LIST_DIR}/exchanger.hpp
+  PUBLIC ${CMAKE_CURRENT_LIST_DIR}/exchanger.cpp
+  PUBLIC ${CMAKE_CURRENT_LIST_DIR}/mpi.hpp
+  PUBLIC ${CMAKE_CURRENT_LIST_DIR}/mpi.cpp
+)
